Make pywin.mfc.afxres canonical and deprecate global afxres
