+++ Pythonwin/pywin/framework/editor/ModuleBrowser.py
@@ -1,13 +1,13 @@
 # ModuleBrowser.py - A view that provides a module browser for an editor document.
 import pyclbr
 
-import afxres
 import commctrl
 import pywin.framework.scriptutils
 import pywin.mfc.docview
 import win32api
 import win32con
 import win32ui
+from pywin.mfc import afxres
 from pywin.tools import browser, hierlist
