+++ Pythonwin/pywin/framework/editor/editor.py
@@ -19,7 +19,6 @@
 
 import re
 
-import afxres
 import regex
 import win32api
 import win32con
@@ -29,6 +28,7 @@ from pywin.framework.editor import (
     GetEditorOption,
     defaultCharacterFormat,
 )
+from pywin.mfc import afxres
 from pywin.mfc.docview import RichEditView as ParentEditorView
 
 from .document import EditorDocumentBase as ParentEditorDocument
