+++ Pythonwin/pywin/scintilla/find.py
@@ -1,10 +1,9 @@
 # find.py - Find and Replace
-import afxres
 import win32api
 import win32con
 import win32ui
 from pywin.framework import scriptutils
-from pywin.mfc import dialog
+from pywin.mfc import afxres, dialog
 
 FOUND_NOTHING = 0
 FOUND_NORMAL = 1
